// src/types/generated/BatchProgressEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TranslationItem } from './TranslationItem';
import type { TranslationStats } from './TranslationStats';

export type BatchProgressEvent = {
  task_id: number;
  processed: number;
  total: number;
  items: Array<TranslationItem>;
  stats: TranslationStats;
};