// src/types/generated/BatchResultWithTaskId.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TranslationItem } from './TranslationItem';
import type { TranslationStats } from './TranslationStats';

export type BatchResultWithTaskId = {
  task_id: number;
  items: Array<TranslationItem>;
  cancelled: boolean;
  stats: TranslationStats;
};